+++ example/dart_days_example.dart
@@ -6,4 +6,5 @@ void main() {
   print('Times Left: ${dartDays.timesLeft(to: DateTime.parse('2020-01-02'))}');
   print('Get number of days in current month: ${DartDays.numberOfDaysInThisMonth()}');
   print('Days Name of week: ${DartDays.daysNameOfWeek().join(", ")}');
+  print('Name of this month ${DartDays.nameOfMonth(1)}');
 }